+++ glassure/gui/widgets/control/composition.py
@@ -1,7 +1,9 @@
 # -*- coding: utf-8 -*-
+from qtpy import QtCore, QtWidgets
 
-from qtpy import QtCore, QtGui, QtWidgets
+from ..custom import FloatLineEdit, CommaDoubleValidator
 from ....core.scattering_factors import calculators, sources
+
 
 Signal = QtCore.Signal
 
@@ -17,6 +19,7 @@ class CompositionWidget(QtWidgets.QWidget):
         self._style_widgets()
 
     def _create_widgets(self):
+
         self.source_lbl = QtWidgets.QLabel("Source:")
         self.source_cb = QtWidgets.QComboBox()
         self.source_cb.addItems(sources)
@@ -26,11 +29,12 @@ class CompositionWidget(QtWidgets.QWidget):
         self.delete_element_btn = QtWidgets.QPushButton("Delete")
 
         self.density_lbl = QtWidgets.QLabel("Density:")
-        self.density_txt = QtWidgets.QLineEdit("2.2")
+        self.density_txt = FloatLineEdit("2.2")
         self.density_atomic_units_lbl = QtWidgets.QLabel("")
 
         self.composition_tw = QtWidgets.QTableWidget()
-        self.composition_tw.cellChanged.connect(self.emit_composition_changed_signal)
+        self.composition_tw.cellChanged.connect(
+            self.emit_composition_changed_signal)
 
     def _create_layout(self):
         self.main_layout = QtWidgets.QVBoxLayout()
@@ -58,18 +62,22 @@ class CompositionWidget(QtWidgets.QWidget):
         self.main_layout.addLayout(self.button_layout)
         self.main_layout.addWidget(self.composition_tw)
         self.main_layout.addLayout(self.density_layout)
-
         self.setLayout(self.main_layout)
 
     def _style_widgets(self):
-        self.source_lbl.setAlignment(QtCore.Qt.AlignVCenter | QtCore.Qt.AlignRight)
 
-        self.density_lbl.setAlignment(QtCore.Qt.AlignVCenter | QtCore.Qt.AlignRight)
-        self.density_atomic_units_lbl.setAlignment(QtCore.Qt.AlignVCenter | QtCore.Qt.AlignRight)
+        self.source_lbl.setAlignment(
+            QtCore.Qt.AlignVCenter | QtCore.Qt.AlignRight)
 
-        self.density_txt.setAlignment(QtCore.Qt.AlignRight)
-        self.density_txt.setValidator(QtGui.QDoubleValidator())
-        self.density_txt.editingFinished.connect(self.emit_composition_changed_signal)
+        self.density_lbl.setAlignment(
+            QtCore.Qt.AlignVCenter | QtCore.Qt.AlignRight)
+
+        self.density_atomic_units_lbl.setAlignment(
+            QtCore.Qt.AlignVCenter | QtCore.Qt.AlignRight)
+
+        self.density_txt.editingFinished.connect(
+            self.emit_composition_changed_signal)
+
         self.density_txt.setMaximumWidth(100)
 
         self.composition_tw.setColumnCount(2)
@@ -83,22 +91,30 @@ class CompositionWidget(QtWidgets.QWidget):
         current_rows = self.composition_tw.rowCount()
         self.composition_tw.setRowCount(current_rows + 1)
         self.composition_tw.blockSignals(True)
+
         element_cb = QtWidgets.QComboBox(self)
         element_cb.setStyle(QtWidgets.QStyleFactory.create('cleanlooks'))
 
-        for ind, ele in enumerate(calculators[self.source_cb.currentText()].elements):
+        for ind, ele in enumerate(calculators[self.source_cb.currentText()].
+                                  elements):
             element_cb.insertItem(ind, ele)
 
         if element is not None:
             element_cb.setCurrentIndex(element_cb.findText(element))
+
         self.composition_tw.setCellWidget(current_rows, 0, element_cb)
 
-        element_cb.currentIndexChanged.connect(self.emit_composition_changed_signal)
+        element_cb.currentIndexChanged.connect(
+            self.emit_composition_changed_signal)
+
         if value is not None:
             value_item = QtWidgets.QTableWidgetItem(str(value))
         else:
             value_item = QtWidgets.QTableWidgetItem(str(1))
-        value_item.setTextAlignment(QtCore.Qt.AlignRight | QtCore.Qt.AlignVCenter)
+
+        value_item.setTextAlignment(
+            QtCore.Qt.AlignRight | QtCore.Qt.AlignVCenter)
+
         self.composition_tw.setItem(current_rows, 1, value_item)
         self.composition_tw.blockSignals(False)
         self.emit_composition_changed_signal()
@@ -120,10 +136,13 @@ class CompositionWidget(QtWidgets.QWidget):
 
     def get_composition(self):
         composition = {}
+
         for row_ind in range(self.composition_tw.rowCount()):
             cb_item = self.composition_tw.cellWidget(row_ind, 0)
             value_item = self.composition_tw.item(row_ind, 1)
-            composition[str(cb_item.currentText())] = float(str(value_item.text()))
+            composition[str(cb_item.currentText())] = float(
+                str(value_item.text().replace(',', '.')))
+
         return composition
 
     def get_sf_source(self):
@@ -133,10 +152,11 @@ class CompositionWidget(QtWidgets.QWidget):
         self.source_cb.setCurrentIndex(self.source_cb.findText(source))
 
     def get_density(self):
-        return float(str(self.density_txt.text()).replace(",", "."))
+        return self.density_txt.value()
 
     def emit_composition_changed_signal(self):
-        self.composition_changed.emit(self.get_composition(), self.get_density())
+        self.composition_changed.emit(
+            self.get_composition(), self.get_density())
 
 
 class TextDoubleDelegate(QtWidgets.QStyledItemDelegate):
@@ -146,7 +166,7 @@ class TextDoubleDelegate(QtWidgets.QStyledItemDelegate):
     def createEditor(self, parent, _, model):
         self.editor = QtWidgets.QLineEdit(parent)
         self.editor.setFrame(False)
-        self.editor.setValidator(QtGui.QDoubleValidator())
+        self.editor.setValidator(CommaDoubleValidator())
         self.editor.setAlignment(QtCore.Qt.AlignRight | QtCore.Qt.AlignVCenter)
         return self.editor