+++ glassure/gui/widgets/control/optimization.py
@@ -1,7 +1,8 @@
 # -*- coding: utf-8 -*-
+from qtpy import QtCore, QtWidgets
 
-from qtpy import QtCore, QtGui, QtWidgets
-from ..custom import NumberTextField, LabelAlignRight, HorizontalLine
+from ..custom import FloatLineEdit, IntegerLineEdit, LabelAlignRight
+from ..custom.lines import HorizontalLine
 
 Signal = QtCore.Signal
 
@@ -24,10 +25,10 @@ class OptimizationWidget(QtWidgets.QWidget):
         self.activate_cb = QtWidgets.QCheckBox("activate")
 
         self.r_cutoff_lbl = LabelAlignRight('r cutoff:')
-        self.r_cutoff_txt = NumberTextField('1.4')
+        self.r_cutoff_txt = FloatLineEdit('1.4')
 
         self.optimize_iterations_lbl = LabelAlignRight("Iterations:")
-        self.optimize_iterations_txt = NumberTextField('5')
+        self.optimize_iterations_txt = IntegerLineEdit('5')
 
         self.attenuation_factor_lbl = LabelAlignRight("Attenuation:")
         self.attenuation_factor_sb = QtWidgets.QSpinBox()
@@ -38,9 +39,6 @@ class OptimizationWidget(QtWidgets.QWidget):
 
         self.r_cutoff_txt.setMaximumWidth(80)
         self.optimize_iterations_txt.setMaximumWidth(80)
-
-        self.r_cutoff_txt.setValidator(QtGui.QDoubleValidator())
-        self.optimize_iterations_txt.setValidator(QtGui.QIntValidator())
 
         self.attenuation_factor_sb.setRange(1, 1000)
         self.attenuation_factor_sb.setSingleStep(1)
@@ -80,11 +78,15 @@ class OptimizationWidget(QtWidgets.QWidget):
 
     def create_signals(self):
         self.activate_cb.stateChanged.connect(self.param_widget.setVisible)
-        self.activate_cb.stateChanged.connect(self.emit_calculation_changed_signal)
+        self.activate_cb.stateChanged.connect(
+            self.emit_calculation_changed_signal)
 
-        self.r_cutoff_txt.editingFinished.connect(self.emit_calculation_changed_signal)
-        self.optimize_iterations_txt.editingFinished.connect(self.emit_calculation_changed_signal)
-        self.attenuation_factor_sb.valueChanged.connect(self.emit_calculation_changed_signal)
+        self.r_cutoff_txt.editingFinished.connect(
+            self.emit_calculation_changed_signal)
+        self.optimize_iterations_txt.editingFinished.connect(
+            self.emit_calculation_changed_signal)
+        self.attenuation_factor_sb.valueChanged.connect(
+            self.emit_calculation_changed_signal)
 
     def emit_calculation_changed_signal(self):
         if self.r_cutoff_txt.isModified():